+++ src/PlaywrightRunner/Models/FlowSpecification.cs
@@ -2,6 +2,7 @@ namespace PlaywrightRunner.Models;
 
 public static class FlowSpecification
 {
+    public const int LegacyVersion = 1;
     public const int CurrentVersion = 2;
-    public const int MinimumSupportedVersion = 2;
+    public const int MinimumSupportedVersion = LegacyVersion;
 }